requirements.txt
# local package
# -e .

# external requirements
click
coverage
black
python-dotenv>=0.5.1
dm-haiku
optax
jupyter
gym
distrax --install-option="--no-deps"
joblib
mujoco_py
